Code analysis fix, methods renamed and typo bug fix

+++ GrEmit/Utils/HackHelpers.cs
@@ -17,13 +17,15 @@ namespace GrEmit.Utils
         //NOTE MetadataToken's reassigned on compilation !!! use inside appdomain
         public static ulong GetMemberUniqueToken(MemberInfo mi)
         {
-            return ((ulong)mi.Module.MetadataToken) << 32 | (ulong)mi.MetadataToken;
+            //CS0675 fix
+            return ((ulong)(long)mi.Module.MetadataToken) << 32 | (ulong)(long)mi.MetadataToken;
         }
 
         //BUG может быть одинаков для типов из разных сборок
         public static ulong GetTypeUniqueToken(Type type)
         {
-            return ((ulong)type.Module.MetadataToken) << 32 | (ulong)type.MetadataToken;
+            //CS0675 fix
+            return ((ulong)(long)type.Module.MetadataToken) << 32 | (ulong)(long)type.MetadataToken;
         }
 
         public static ConstructorInfo GetObjectConstruction<T>(Expression<Func<T>> constructorCall,
